+++ saleor/graphql/meta/tests/test_meta_mutations.py
@@ -383,7 +383,9 @@ def test_add_public_metadata_for_checkout(api_client, checkout):
 
     # then
     assert item_contains_proper_public_metadata(
-        response["data"]["updateMetadata"]["item"], checkout, checkout_id
+        response["data"]["updateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -413,7 +415,9 @@ def test_add_public_metadata_for_checkout_by_token(api_client, checkout):
 
     # then
     assert item_contains_proper_public_metadata(
-        response["data"]["updateMetadata"]["item"], checkout, checkout_id
+        response["data"]["updateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -911,8 +915,8 @@ def test_add_public_metadata_for_menu_item(
 
 def test_update_public_metadata_for_item(api_client, checkout):
     # given
-    checkout.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
-    checkout.save(update_fields=["metadata"])
+    checkout.metadata_storage.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    checkout.metadata_storage.save(update_fields=["metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -923,7 +927,7 @@ def test_update_public_metadata_for_item(api_client, checkout):
     # then
     assert item_contains_proper_public_metadata(
         response["data"]["updateMetadata"]["item"],
-        checkout,
+        checkout.metadata_storage,
         checkout_id,
         value="NewMetaValue",
     )
@@ -971,8 +975,8 @@ def test_update_public_metadata_for_order_line(api_client, order_line):
 
 @pytest.mark.django_db(transaction=True)
 def test_update_public_metadata_for_item_on_deleted_instance(api_client, checkout):
-    checkout.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
-    checkout.save(update_fields=["metadata"])
+    checkout.metadata_storage.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    checkout.metadata_storage.save(update_fields=["metadata"])
 
     def delete_checkout_object(*args, **kwargs):
         with transaction.atomic():
@@ -1165,7 +1169,6 @@ def execute_clear_public_metadata_for_item(
         "id": item_id,
         "keys": [key],
     }
-
     response = client.post_graphql(
         DELETE_PUBLIC_METADATA_MUTATION % item_type,
         variables,
@@ -1365,8 +1368,8 @@ def test_delete_public_metadata_for_myself_as_staff(staff_api_client):
 
 def test_delete_public_metadata_for_checkout(api_client, checkout):
     # given
-    checkout.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
-    checkout.save(update_fields=["metadata"])
+    checkout.metadata_storage.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    checkout.metadata_storage.save(update_fields=["metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -1376,14 +1379,16 @@ def test_delete_public_metadata_for_checkout(api_client, checkout):
 
     # then
     assert item_without_public_metadata(
-        response["data"]["deleteMetadata"]["item"], checkout, checkout_id
+        response["data"]["deleteMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
 def test_delete_public_metadata_for_checkout_by_token(api_client, checkout):
     # given
-    checkout.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
-    checkout.save(update_fields=["metadata"])
+    checkout.metadata_storage.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    checkout.metadata_storage.save(update_fields=["metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -1393,7 +1398,9 @@ def test_delete_public_metadata_for_checkout_by_token(api_client, checkout):
 
     # then
     assert item_without_public_metadata(
-        response["data"]["deleteMetadata"]["item"], checkout, checkout_id
+        response["data"]["deleteMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -1865,8 +1872,8 @@ def test_delete_public_metadata_for_item_without_meta(api_client, address):
 
 def test_delete_public_metadata_for_not_exist_key(api_client, checkout):
     # given
-    checkout.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
-    checkout.save(update_fields=["metadata"])
+    checkout.metadata_storage.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    checkout.metadata_storage.save(update_fields=["metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -1876,16 +1883,18 @@ def test_delete_public_metadata_for_not_exist_key(api_client, checkout):
 
     # then
     assert item_contains_proper_public_metadata(
-        response["data"]["deleteMetadata"]["item"], checkout, checkout_id
+        response["data"]["deleteMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
 def test_delete_public_metadata_for_one_key(api_client, checkout):
     # given
-    checkout.store_value_in_metadata(
+    checkout.metadata_storage.store_value_in_metadata(
         {PUBLIC_KEY: PUBLIC_VALUE, "to_clear": PUBLIC_VALUE},
     )
-    checkout.save(update_fields=["metadata"])
+    checkout.metadata_storage.save(update_fields=["metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -1895,11 +1904,13 @@ def test_delete_public_metadata_for_one_key(api_client, checkout):
 
     # then
     assert item_contains_proper_public_metadata(
-        response["data"]["deleteMetadata"]["item"], checkout, checkout_id
+        response["data"]["deleteMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
     assert item_without_public_metadata(
         response["data"]["deleteMetadata"]["item"],
-        checkout,
+        checkout.metadata_storage,
         checkout_id,
         key="to_clear",
     )
@@ -2217,7 +2228,9 @@ def test_add_private_metadata_for_checkout(
 
     # then
     assert item_contains_proper_private_metadata(
-        response["data"]["updatePrivateMetadata"]["item"], checkout, checkout_id
+        response["data"]["updatePrivateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -2253,7 +2266,9 @@ def test_add_private_metadata_for_checkout_by_token(
 
     # then
     assert item_contains_proper_private_metadata(
-        response["data"]["updatePrivateMetadata"]["item"], checkout, checkout_id
+        response["data"]["updatePrivateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -2720,8 +2735,10 @@ def test_update_private_metadata_for_item(
     staff_api_client, checkout, permission_manage_checkouts
 ):
     # given
-    checkout.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_KEY})
-    checkout.save(update_fields=["private_metadata"])
+    checkout.metadata_storage.store_value_in_private_metadata(
+        {PRIVATE_KEY: PRIVATE_KEY}
+    )
+    checkout.metadata_storage.save(update_fields=["private_metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -2736,7 +2753,7 @@ def test_update_private_metadata_for_item(
     # then
     assert item_contains_proper_private_metadata(
         response["data"]["updatePrivateMetadata"]["item"],
-        checkout,
+        checkout.metadata_storage,
         checkout_id,
         value="NewMetaValue",
     )
@@ -3210,8 +3227,10 @@ def test_delete_private_metadata_for_checkout(
     staff_api_client, checkout, permission_manage_checkouts
 ):
     # given
-    checkout.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
-    checkout.save(update_fields=["private_metadata"])
+    checkout.metadata_storage.store_value_in_private_metadata(
+        {PRIVATE_KEY: PRIVATE_VALUE}
+    )
+    checkout.metadata_storage.save(update_fields=["private_metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -3221,7 +3240,9 @@ def test_delete_private_metadata_for_checkout(
 
     # then
     assert item_without_private_metadata(
-        response["data"]["deletePrivateMetadata"]["item"], checkout, checkout_id
+        response["data"]["deletePrivateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -3229,8 +3250,10 @@ def test_delete_private_metadata_for_checkout_by_token(
     staff_api_client, checkout, permission_manage_checkouts
 ):
     # given
-    checkout.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
-    checkout.save(update_fields=["private_metadata"])
+    checkout.metadata_storage.store_value_in_private_metadata(
+        {PRIVATE_KEY: PRIVATE_VALUE}
+    )
+    checkout.metadata_storage.save(update_fields=["private_metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -3240,7 +3263,9 @@ def test_delete_private_metadata_for_checkout_by_token(
 
     # then
     assert item_without_private_metadata(
-        response["data"]["deletePrivateMetadata"]["item"], checkout, checkout_id
+        response["data"]["deletePrivateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -3736,8 +3761,10 @@ def test_delete_private_metadata_for_not_exist_key(
     staff_api_client, checkout, permission_manage_checkouts
 ):
     # given
-    checkout.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
-    checkout.save(update_fields=["private_metadata"])
+    checkout.metadata_storage.store_value_in_private_metadata(
+        {PRIVATE_KEY: PRIVATE_VALUE}
+    )
+    checkout.metadata_storage.save(update_fields=["private_metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -3751,7 +3778,9 @@ def test_delete_private_metadata_for_not_exist_key(
 
     # then
     assert item_contains_proper_private_metadata(
-        response["data"]["deletePrivateMetadata"]["item"], checkout, checkout_id
+        response["data"]["deletePrivateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
 
 
@@ -3759,10 +3788,10 @@ def test_delete_private_metadata_for_one_key(
     staff_api_client, checkout, permission_manage_checkouts
 ):
     # given
-    checkout.store_value_in_private_metadata(
+    checkout.metadata_storage.store_value_in_private_metadata(
         {PRIVATE_KEY: PRIVATE_VALUE, "to_clear": PRIVATE_VALUE},
     )
-    checkout.save(update_fields=["private_metadata"])
+    checkout.metadata_storage.save(update_fields=["private_metadata"])
     checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     # when
@@ -3776,11 +3805,13 @@ def test_delete_private_metadata_for_one_key(
 
     # then
     assert item_contains_proper_private_metadata(
-        response["data"]["deletePrivateMetadata"]["item"], checkout, checkout_id
+        response["data"]["deletePrivateMetadata"]["item"],
+        checkout.metadata_storage,
+        checkout_id,
     )
     assert item_without_private_metadata(
         response["data"]["deletePrivateMetadata"]["item"],
-        checkout,
+        checkout.metadata_storage,
         checkout_id,
         key="to_clear",
     )